Use navigator.userAgentData in Chrome version page
navigator.userAgent will be deprecated in Chrome 101, so this CL makes
the JS code use navigator.userAgentData to keep retrieving the browser
version, while keeping the deprecated way as a fallback.

In order to simplify the logic now the page only shows the version for
Chrome, and doesn't work with other browsers.

Also, due to the complexity of parsing the Chrome iOS user agent string,
this OS isn't supported yet.

Fixed: misc:43
Change-Id: I23e7221318013afb3b1bba738e9c07851ea1692f